requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
WTForms==3.1.1
Flask-Mail==0.9.1
Flask-CORS==4.0.0
python-dotenv==1.0.0
email-validator==2.0.0
gunicorn==21.2.0
psycopg2-binary==2.9.9